requirements.txt
torch
transformers
peft
accelerate
datasets
trl
bitsandbytes
python-dotenv
scipy
pandas 
openpyxl 
tabulate
